requirements.txt
absl-py==1.0.0
albumentations==1.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
fvcore==0.1.5.post20220414
google-auth==2.6.5
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
imageio==2.16.2
importlib-metadata==4.11.3
iopath==0.1.9
joblib==1.1.0
kornia==0.6.4
Markdown==3.3.6
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==9.1.0
portalocker==2.4.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.4.8
timm
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
yacs==0.1.8
zipp==3.8.0
